Added in Topology Commands & State Tracking, ABLE TO RENDER AGAIN!

+++ ProjectGE/src/ProjectGE/Rendering/DirectX12/Resources/DirectX12IndexBuffer.h
@@ -7,13 +7,13 @@ namespace ProjectGE {
 	public:
 		DirectX12IndexBuffer() = default;
 
-		DirectX12IndexBuffer(unsigned int* buffer, unsigned int count);
+		DirectX12IndexBuffer(unsigned short* buffer, unsigned int count);
 
 		D3D12_INDEX_BUFFER_VIEW GenerateView();
 
 		UINT GetCount() override;
 	private:
-		Scope<DirectX12Buffer<unsigned int>> m_Buffer;
+		Scope<DirectX12Buffer<unsigned short>> m_Buffer;
 		UINT m_Count;
 	};
 };
